Improved error detection in some cases
We now wait for a maximum of 10 messages, or 10 seconds.. or rather
it is a combination, it waits for a new message for a maximum of one
second and does this a maximum of 10 times.

Some (most) functions should now throw an exception for some (most?)
of the things that can go wrong with it.

Still things to do:
- Add more error detection
- Improve exceptions
- Recover from errors