resources: Update udev rules
Rename the rules file so its ordering in the set of rules is more
tightly defined. The range 4* is typically used for USB serial devices
(for example,
https://github.com/Nitrokey/libnitrokey/blob/master/data/41-nitrokey.rules).

Use the new `uaccess` tag to give the user access to the device, rather
than hard-coding world-accessible permissions. (See
https://github.com/systemd/systemd/issues/4288.)

Drop the `/dev/ttyANT*` symlink, which was unused in the code and just
cluttered up the devfs. Instead, set the `ID_ANT_DEVICE` variable on the
device. The code in openant could eventually key on that to find USB
devices to drive (currently it manually matches by vendor and product).

Signed-off-by: Philip Withnall <philip@tecnocode.co.uk>